clean up dependencies
* setup.py will only try to install pynq if you're on a supported CPU arch
* move QickConfig to qick_asm.py, which makes sense thematically and allows it to be used on non-Zynq machines (which is the whole point)
